Refactor code to use web workers
Remove the dynamic import of modules and replace it
with web workers.

This solves the problem with slow functions that blocks
the user from doing anything until the function has finished.

With web workers the calculations are moved from the main thread
to a separate thread. This makes it possible for the user to still
use the UI and click on other buttons to calculate other problems
simultaneously.